lighting処理

+++ Engine/3D/ObjModelObject.h
@@ -46,7 +46,8 @@ public: // サブクラス
 		Vector3 cameraPos;		//カメラ座標
 		float pad1;
 		Vector4 color;
-		int mask;		//ライティングを当てるか
+		//float pad2;
+		float mask;		//ライティングを当てるか
 	};
 	struct ConstBufferDataB1
 	{
@@ -84,7 +85,7 @@ public: // 静的メンバ関数
 	//Setter
 	inline void SetModel(ObjModelManager* model){this->model = model;}
 	inline void SetColor(Vector4 _color)	{this->color = _color;}
-	inline void SetLightMask(int maskValue) {this->mask = maskValue;}
+	inline void SetLightMask(float maskValue) {this->mask = maskValue;}
 
 private: // 静的メンバ変数
 	static CommonObj* common;
@@ -114,9 +115,9 @@ private: // メンバ変数
 	ComPtr<ID3D12Resource> constBuffB1; // 定数バッファ
 
 	//色
-	Vector4 color = {1.0f,1.0f,1.0f, 1.0f};
+	Vector4 color = {1.0f,1.0f,1.0f,1.0f};
 
 	//ライティングマスク
-	int mask = 0;
+	float mask = 0.0f;
 };
